Revert "cache"
This reverts commit 46a75a7a791b890f289d9b4e867a1f444dc80016.
