+++ browser/views/fileManager/styles/fileDescriptor.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const { AWAITING_DOWNLOAD, DOWNLOADING } = require('../../../../lib/constants/stateManagement')
-const { fonts, colorSelector } = require('styleUtils')
+const { AWAITING_DOWNLOAD } = require('../../../../lib/constants/stateManagement')
+const { fonts } = require('styleUtils')
 const { css } = require('css')
 
 module.exports = {
@@ -20,29 +20,16 @@ module.exports = {
   container: css`
     :host {
       display: flex;
+      height: 75px;
       width: 100%;
     }
   `,
 
-  exclamation: css`
-    :host {
-      color: red;
-      font-size: 20px;
-    }
-  `,
-
-  hamburgerHolder: css`
+  iconHolder: css`
     :host {
       display: flex;
-      align-items: center;
-      width: 9%;
-    }
-  `,
-
-  hamburger: css`
-    :host {
-      height: 8px;
-      width: 40%;
+      justify-content: left;
+      width: 12%;
     }
   `,
 
@@ -57,31 +44,18 @@ module.exports = {
   nameHolder: css`
     :host {
       display: flex;
-      font-family: ${fonts.bold} !important;
-      font-weight: bold;
-      font-size: 17px;
-      margin-bottom: 2px;
+      font-family: ${fonts.bold};
+      font-size: 16px;
       width: 75%;
     }
   `,
 
   sizeHolder(status) {
-    let color
-    switch (status) {
-      case AWAITING_DOWNLOAD:
-        color = 'grey'
-        break
-      case DOWNLOADING:
-        color = 'red'
-        break
-      default:
-        color = 'blue'
-    }
     return css`
       :host {
-        color: ${colorSelector(color)};
+        color: ${status === AWAITING_DOWNLOAD ? 'grey' : 'black' };
         font-family: ${fonts.light};
-        font-size: 12px;
+        font-size: 13px;
       }
     `
   },
