+++ lib/constants/stateManagement.js
@@ -28,7 +28,6 @@ module.exports = {
   LOGOUT: 'LOGOUT',
   NEW_STATE: 'NEW_STATE',
   OPEN_MODAL: 'OPEN_MODAL',
-  OUT_OF_SYNC: 'OUT_OF_SYNC',
   PROMPT_PURCHASE: 'PROMPT_PURCHASE',
   PUBLISH: 'PUBLISH',
   PUBLISHED: 'PUBLISHED',
@@ -40,7 +39,7 @@ module.exports = {
   REGISTER: 'REGISTER',
   REGISTERED: 'REGISTERED',
   REGISTERING: 'REGISTERING',
-  UPDATE_AVAILABLE: 'UPDATE_AVAILABLE',
+  UPLOAD_COMPLETE: 'UPLOAD_COMPLETE',
   UPDATE_FILE: 'UPDATE_FILE',
   UPDATED_FILE: 'UPDATED_FILE',
   UPDATING_FILE: 'UPDATING_FILE',
